+++ react/components/EXPERIMENTAL_Select/index.js
@@ -27,25 +27,7 @@ class Select extends Component {
   constructor(props) {
     super(props)
 
-    this.state = {
-      searchTerm: undefined,
-    }
-
     this.inputId = `react-select-input-${uuid()}`
-  }
-
-  componentDidUpdate(prevProps, prevState) {
-    const { searchTerm } = this.state
-    const { searchTerm: prevSearchTerm } = prevState
-    const { loading } = this.props
-    const { loading: prevLoading } = prevProps
-
-    if (
-      searchTerm !== prevSearchTerm ||
-      (searchTerm && loading !== prevLoading)
-    ) {
-      document.getElementById(this.inputId).focus()
-    }
   }
 
   render() {
@@ -80,20 +62,12 @@ class Select extends Component {
       ref: forwardedRef,
       autoFocus,
       className: `pointer bw1 ${getFontClassNameFromSize(size)}`,
+      errorMessage,
+      size,
       components: {
         ClearIndicator,
-        Control: function Control(props) {
-          return (
-            <ControlComponent
-              errorMessage={errorMessage}
-              size={size}
-              {...props}
-            />
-          )
-        },
-        DropdownIndicator: function DropdownIndicator(props) {
-          return <DropdownIndicatorComponent size={size} {...props} />
-        },
+        Control: ControlComponent,
+        DropdownIndicator: DropdownIndicatorComponent,
         IndicatorSeparator: () => null,
         MultiValueRemove,
         Placeholder,
@@ -110,9 +84,6 @@ class Select extends Component {
       noOptionsMessage,
       inputId: this.inputId,
       onInputChange: (value, { action }) => {
-        this.setState({
-          searchTerm: value,
-        })
         if (
           action === 'input-change' &&
           typeof onSearchInputChange === 'function'